+++ collector/src/smokeProduction/java/com/aspix2k/affected/collector/smoke/ProductionService.java
@@ -1,5 +1,0 @@
-package com.aspix2k.affected.collector.smoke;
-
-public interface ProductionService {
-    int value();
-}