+++ collector/src/smokeProduction/java/com/aspix2k/affected/collector/smoke/ProductionServiceImpl.java
@@ -1,8 +1,0 @@
-package com.aspix2k.affected.collector.smoke;
-
-public final class ProductionServiceImpl implements ProductionService {
-    @Override
-    public int value() {
-        return 2;
-    }
-}