+++ conformance/cli-fixtures/dotnet/MSTest.Tests/AlphaTest.cs
@@ -1,8 +1,0 @@
-namespace MSTest.Tests;
-
-[Microsoft.VisualStudio.TestTools.UnitTesting.TestClass]
-public class AlphaTest
-{
-    [Microsoft.VisualStudio.TestTools.UnitTesting.TestMethod]
-    public void Passes() => Microsoft.VisualStudio.TestTools.UnitTesting.Assert.AreEqual(1, Alpha.AlphaValue.Get());
-}