+++ conformance/cli-fixtures/go/alpha/alpha_test.go
@@ -1,9 +1,0 @@
-package alpha
-
-import "testing"
-
-func TestValue(t *testing.T) {
-	if Value() != 1 {
-		t.Fatal("unexpected value")
-	}
-}