+++ conformance/cli-fixtures/go/beta/beta_test.go
@@ -1,9 +1,0 @@
-package beta
-
-import "testing"
-
-func TestValue(t *testing.T) {
-	if Value() != 2 {
-		t.Fatal("unexpected value")
-	}
-}